Update README.md
NIDM (National Institute Of Digital Marketing) Bangalore Is One Of The Leading & best Digital Marketing Institute In Bangalore, India And We Have Brand Value For The Quality Of Education Which We Provide. Our Curriculum/ Courses Are Designed with Practical knowledge are Fully For Job Orientation Bases. We have the best curriculum, trainers and unlimited practical hours on live project